// patchnotes-web/src/api/generated/model/packageDetailGroupDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * PatchNotes.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PackageDetailReleaseDto } from './packageDetailReleaseDto';

export interface PackageDetailGroupDto {
  majorVersion?: number;
  isPrerelease?: boolean;
  versionRange: string;
  /** @nullable */
  summary?: string | null;
  releaseCount?: number;
  lastUpdated?: string;
  releases: PackageDetailReleaseDto[];
}