// patchnotes-web/src/api/generated/model/packageDetailReleaseDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * PatchNotes.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PackageDetailReleaseDto {
  id: string;
  tag: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  publishedAt?: string;
}